Adapt puppet for host-based vs containers install
Make Puppet an empty meta-package that inlcudes everything from
puppet-systemd-initd and puppet-headless subpackages. The former
consumes the bits required to manipulate to systemd units/tmpfiles.d,
including buildtime steps and initd configuration as well. It is
targeted for the host configuration mainly and should not be used for
containers images building normally.

While the puppet-headless subpackage may be used instead of the
main puppet package, when installing it in containers images so there
is no need to create and run any service units and the like.

Resolves: rhbz#1654672

Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>